handle naviagte to home screen when user have been authorized
